+++ src/prefect/cli/dev.py
@@ -21,7 +21,6 @@ from prefect.cli._types import PrefectTyper, SettingsOption
 from prefect.cli._utilities import exit_with_error, exit_with_success
 from prefect.cli.agent import start as start_agent
 from prefect.cli.root import app
-from prefect.docker import get_prefect_image_name, python_version_minor
 from prefect.server.api.server import create_app
 from prefect.settings import (
     PREFECT_API_URL,
@@ -30,6 +29,7 @@ from prefect.settings import (
     PREFECT_SERVER_API_HOST,
     PREFECT_SERVER_API_PORT,
 )
+from prefect.utilities.dockerutils import get_prefect_image_name, python_version_minor
 from prefect.utilities.filesystem import tmpchdir
 from prefect.utilities.processutils import run_process
